Updates to factories - this will break tests (more than they already are)
- Moved `program_with_requirements` out of courses/models_test.py and into factories.py (seemed more appropriate)
- Fixed an actual bug in `get_program_certificate_by_enrollment` (missing not operator)
- Fixed tests for `utils_test`
