+++ courses/utils_test.py
@@ -2,6 +2,8 @@
 """
 Tests for utils
 """
+import pytest
+
 from courses.factories import (
     CourseFactory,
     CourseRunEnrollmentFactory,
@@ -9,26 +11,34 @@ from courses.factories import (
     ProgramCertificateFactory,
     ProgramEnrollmentFactory,
     ProgramFactory,
+    program_with_requirements,
 )
-from courses.utils import (
-    get_program_certificate_by_enrollment,
-)
+from courses.utils import get_program_certificate_by_enrollment
 
 
-def test_get_program_certificate_by_enrollment(user, program):
+def test_get_program_certificate_by_enrollment(user, program_with_requirements):
     """
     Test that get_program_certificate_by_enrollment returns a program certificate
     """
-    course = CourseFactory.create(program=program)
+    course = program_with_requirements.program.courses[0][0]
     course_run = CourseRunFactory.create(course=course)
 
     course_enrollment = CourseRunEnrollmentFactory.create(user=user, run=course_run)
-    program_enrollment = ProgramEnrollmentFactory.create(user=user, program=program)
+    program_enrollment = ProgramEnrollmentFactory.create(
+        user=user, program=program_with_requirements.program
+    )
 
-    program_certificate = ProgramCertificateFactory.create(program=program, user=user)
+    program_certificate = ProgramCertificateFactory.create(
+        program=program_with_requirements.program, user=user
+    )
+
+    program_with_requirements.program.refresh_from_db()
 
     assert (
-        get_program_certificate_by_enrollment(course_enrollment) == program_certificate
+        get_program_certificate_by_enrollment(
+            course_enrollment, program_with_requirements.program
+        )
+        == program_certificate
     )
     assert (
         get_program_certificate_by_enrollment(program_enrollment) == program_certificate
@@ -39,7 +49,7 @@ def test_get_program_certificate_by_enrollment_program_does_not_exist(user):
     """
     Test that get_program_certificate_by_enrollment returns None if course has no program
     """
-    course = CourseFactory.create(program=None)
+    course = CourseFactory.create()
     course_run = CourseRunFactory.create(course=course)
 
     course_enrollment = CourseRunEnrollmentFactory.create(user=user, run=course_run)
@@ -47,12 +57,17 @@ def test_get_program_certificate_by_enrollment_program_does_not_exist(user):
     assert get_program_certificate_by_enrollment(course_enrollment) == None
 
 
-def test_get_program_certificate_by_enrollment_program_page_does_not_exist(user):
+def test_get_program_certificate_by_enrollment_program_page_does_not_exist(
+    user, program_with_requirements
+):
     """
     Test that get_program_certificate_by_enrollment returns None if program page does not exist
     """
-    course = CourseFactory.create()
-    program = course.program
+    program = program_with_requirements.program
+
+    program.page.delete()
+
+    course = program_with_requirements.program.courses[0][0]
     course_run = CourseRunFactory.create(course=course)
 
     course_enrollment = CourseRunEnrollmentFactory.create(user=user, run=course_run)
@@ -61,23 +76,28 @@ def test_get_program_certificate_by_enrollment_program_page_does_not_exist(user)
     program_certificate = ProgramCertificateFactory.create(program=program, user=user)
 
     assert (
-        get_program_certificate_by_enrollment(course_enrollment) != program_certificate
+        get_program_certificate_by_enrollment(course_enrollment, program)
+        != program_certificate
     )
     assert (
         get_program_certificate_by_enrollment(program_enrollment) != program_certificate
     )
-    assert get_program_certificate_by_enrollment(course_enrollment) == None
+    assert get_program_certificate_by_enrollment(course_enrollment, program) == None
     assert get_program_certificate_by_enrollment(program_enrollment) == None
 
 
 def test_get_program_certificate_by_enrollment_program_certificate_page_does_not_exist(
-    user,
+    user, program_with_requirements
 ):
     """
     Test that get_program_certificate_by_enrollment returns None if program certificate page does not exist
     """
-    program = ProgramFactory(page__certificate_page=None)
-    course = CourseFactory.create(program=program)
+    program = program_with_requirements.program
+
+    program.page.certificate_page.delete()
+    program.page.delete()
+
+    course = program_with_requirements.program.courses[0][0]
     course_run = CourseRunFactory.create(course=course)
 
     course_enrollment = CourseRunEnrollmentFactory.create(user=user, run=course_run)
